+++ packages/skyline-ts/src/index.ts
@@ -16,7 +16,6 @@ export type {
   SlackCloudConfig,
   SlackConfig,
   SlackDedicatedConfig,
-  SlackLine,
   TerminalConfig,
   WhatsappBusinessCloudConfig,
   WhatsappBusinessConfig,